Revert "refactoring to TemporalType enum"
